+++ sync/grandPA/grandPA_client.py
@@ -1,0 +1,58 @@
+# -*- coding: utf-8 -*-
+
+#   Hello World client in Python
+#   Connects REQ socket to tcp://localhost:5555
+#   Sends "Hello" to server, expects "World" back
+#
+
+import zmq
+import argparse
+import time 
+
+
+# Initialize parser
+parser = argparse.ArgumentParser()
+# initalize zmq socket
+context = zmq.Context()
+
+
+#  Open Socket to talk to server
+print("Connecting to hello world server…")
+socket = context.socket(zmq.PUSH)
+socket.connect("tcp://localhost:5555")
+
+
+
+# this is for input arguments: 
+# Adding optional argument
+parser.add_argument("-x", "--xPosition", help = "x Position")
+parser.add_argument("-y", "--yPosition", help = "y Position")
+parser.add_argument("-a", "--Alpha", help = "Alpha value")
+
+parser.add_argument("-KEY_LEFT", "--KEY_LEFT", help = "KEY_LEFT")
+
+ 
+# Read arguments from command line
+args = parser.parse_args()
+
+
+
+# match arguments and send them to server 
+# how would I write this asynchron, so all 3 arguemtnes can be sent simultaniously??
+if args.xPosition:
+    print("X position % s" % args.xPosition)
+    socket.send(b"X % s" % args.xPosition)
+
+if args.yPosition:
+    print("Y position % s" % args.yPosition)
+    socket.send(b"Y % s" % args.yPosition)
+
+if args.Alpha:
+     print("Alpha value % s" % args.Alpha)
+     socket.send(b"a % s" % args.Alpha)
+
+
+if args.KEY_LEFT:
+     print("KEY_LEFT % s" % args.KEY_LEFT)
+     socket.send(b"K_LEFT")
+